refactor(communities): split CommunityGridCard into folder + sub-parts (VWA-96)
